requirements.txt
requests==2.26.0
python-telegram-bot==13.8.1
google-api-python-client==2.32.0
oauth2client==4.1.3
google-cloud-dialogflow==2.10.0
